// frontend/src/types/database.types.ts
export interface LeadDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  salutation: string | null;
  first_name: string | null;
  last_name: string | null;
  title: string | null;
  department: string | null;
  phone_work: string | null;
  phone_mobile: string | null;
  email1: string | null;
  primary_address_street: string | null;
  primary_address_city: string | null;
  primary_address_state: string | null;
  primary_address_postalcode: string | null;
  primary_address_country: string | null;
  status: string | null;
  status_description: string | null;
  lead_source: string | null;
  lead_source_description: string | null;
  description: string | null;
  account_name: string | null;
  website: string | null;
  ai_score: number | null;
  ai_score_date: Date | string | null;
  ai_insights: Record<string, any> | null;
  ai_next_best_action: string | null;
}

export type LeadCreateRequest = Pick<LeadDB,
  'created_by' | 'modified_user_id' | 'assigned_user_id' | 'salutation' | 'first_name' | 'last_name' | 'title' | 'department' | 'phone_work' | 'phone_mobile' | 'email1' | 'primary_address_street' | 'primary_address_city' | 'primary_address_state' | 'primary_address_postalcode' | 'primary_address_country' | 'status' | 'status_description' | 'lead_source' | 'lead_source_description' | 'description' | 'account_name' | 'website' | 'ai_score' | 'ai_score_date' | 'ai_insights' | 'ai_next_best_action'
>;